Fixing problems from the merge (DOES NOT BUILD)
The merge caused some problems because merge conflict only caught the parts I had changed and not when Grey's stuff changed that I happened to be depending on. I tried to fix that but now something else happened and I figured I might as well save the rest of it?